php: rename credentials to channel_credentials
